丰富拦截器逻辑
