+++ net/grpc/interceptor.go
@@ -6,9 +6,6 @@ package grpc
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
-	"path"
 	"time"
 
 	"github.com/FishGoddess/logit"
@@ -17,23 +14,6 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/metadata"
 )
-
-func serviceMethod(info *grpc.UnaryServerInfo) string {
-	if method := path.Base(info.FullMethod); method != "" {
-		return method
-	}
-
-	return info.FullMethod
-}
-
-func jsonify(v any) string {
-	marshaled, err := json.Marshal(v)
-	if err != nil {
-		return fmt.Sprintf("%+v", v)
-	}
-
-	return string(marshaled)
-}
 
 func Interceptor(timeout time.Duration, resolvers ...RequestResolver) grpc.UnaryServerInterceptor {
 	return func(ctx context.Context, req any, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (resp any, err error) {
